TC1/Docker/PostgreSQLPackage/PostgreSQL-Memcached/app/requirements.txt
flask
psycopg2-binary
pymemcache